LayoutTests/http/tests/site-isolation/inspector/network/cross-origin-iframe-network-loading-finished-expected.txt

Test that cross-origin iframe resource loading lifecycle events fire correctly for both success and failure.


== Running test suite: SiteIsolation.Network.CrossOriginIFrameLoadingLifecycle
-- Running test case: SiteIsolation.Network.CrossOriginIFrame.LoadingDidFinish
PASS: Resource should be created.
PASS: Resource should be marked as finished.
PASS: Resource should not be marked as failed.

-- Running test case: SiteIsolation.Network.CrossOriginIFrame.LoadingDidFail
PASS: Resource should be created.
PASS: Resource should be marked as failed.
